Contains a way to read plateContents from NSBundle and a way to read it from the web. Experiments in jSONSerialization

+++ ZipArchiveLibrary/SSZipArchive/SSZipArchive.m
@@ -82,7 +82,7 @@
 		[delegate zipArchiveWillUnzipArchiveAtPath:path zipInfo:globalInfo];
 	}
 	if ([delegate respondsToSelector:@selector(zipArchiveProgressEvent:total:)]) {
-		[delegate zipArchiveProgressEvent:(NSInteger)currentPosition total:(NSInteger)fileSize];
+		[delegate zipArchiveProgressEvent:(int)currentPosition total:(NSInteger)fileSize];
 	}
     
 	NSInteger currentFileNumber = 0;
